+++ packages/common/pipes/src/mark-by-filter/mark-by-filter.interfaces.ts
@@ -1,4 +1,0 @@
-import { SafeHtml } from '@angular/platform-browser';
-
-export type Value = string | null | undefined;
-export type MarkedValue = SafeHtml | string | null | undefined;