+++ packages/common/pipes/src/mark-by-filter/marked-value.ts
@@ -1,0 +1,3 @@
+import { SafeHtml } from '@angular/platform-browser';
+
+export type MarkedValue = SafeHtml | string | null | undefined;