+++ packages/eslint-config/index.js
@@ -55,9 +55,29 @@ module.exports = {
                 'eslint-plugin-import',
                 '@typescript-eslint',
                 '@angular-eslint/eslint-plugin',
-                'sonarjs'
+                'sonarjs',
+                '@angular-ru/eslint-plugin'
             ],
             rules: {
+                /**
+                 * Angular-RU rules
+                 */
+                '@angular-ru/no-suffix-file': [
+                    'error',
+                    {
+                        fileEndsWithList: [
+                            '.model.ts',
+                            '.models.ts',
+                            '.enum.ts',
+                            '.enums.ts',
+                            '.type.ts',
+                            '.types.ts',
+                            '.interface.ts',
+                            '.interfaces.ts'
+                        ]
+                    }
+                ],
+
                 /**
                  * ESLint Core
                  */
@@ -527,7 +547,7 @@ module.exports = {
                 // 'sonarjs/prefer-default-last': 'error',
                 // 'sonarjs/prefer-promise-shorthand': 'error',
                 // 'sonarjs/prefer-type-guard': 'error',
-                // 'sonarjs/use-type-alias': 'error'
+                // 'sonarjs/use-type-alias': 'error',
             }
         },
         {