+++ packages/ng-table-builder/src/services/filterable/filter-columns-opts.ts
@@ -1,0 +1,9 @@
+import { PlainObjectOf } from '@angular-ru/common/typings';
+
+import { TableFilterType } from './table-filter-type';
+
+export interface FilterColumnsOpts {
+    isEmpty: boolean;
+    values: PlainObjectOf<string>;
+    types: PlainObjectOf<TableFilterType>;
+}