+++ packages/ng-table-builder/src/services/filterable/filter-event.ts
@@ -1,0 +1,6 @@
+import { TableFilterType } from './table-filter-type';
+
+export interface FilterEvent {
+    value: string | null;
+    type: string | TableFilterType | null;
+}