+++ packages/ng-table-builder/src/services/filterable/filter-worker-event.ts
@@ -1,0 +1,6 @@
+import { Fn } from '@angular-ru/common/typings';
+
+export interface FilterWorkerEvent<T> {
+    source: T[];
+    fireSelection: Fn;
+}