+++ packages/tooltip/src/interfaces/tooltip-size.ts
@@ -1,0 +1,1 @@
+export type TooltipSize = 'small' | 'medium';