+++ packages/tooltip/src/tooltip-default-text.interceptor.ts
@@ -1,6 +1,0 @@
-import { Injectable } from '@angular/core';
-
-import { TooltipTextInterceptor } from './tooltip.interfaces';
-
-@Injectable()
-export class TooltipDefaultTextInterceptor implements TooltipTextInterceptor {}